WIP, Motor Totally remade from scratch (again)
Now Motor is separated into different interconnected modules, although 
it is more difficult to use it than before, the purpose is to serve as a 
base library for easier and more direct libraries to the game.

That is, the intent is for the developer to use a library that 
communicates with Motor, but not to use Motor directly.

I'm thinking of changing the name since Motor (Engine in portuguese) may 
not reflect so well what this library does.