Refactoring to be compatible with the version* written in Titan**
In fact, both are not yet compatible, this one needs revision, and the Titan version will need to be lightly rewritten as well.

* https://gitlab.com/Andre-LA/motor-titan
** https://github.com/titan-lang/titan

My goal with this:

A Titan code, when compiled, generates C code that is then compiled to .so, and can be used in Lua (just using require). By rewriting Titan as an experiment, I came up with a more organized result than the current one in Lua.

The version in Titan is faster than in Lua, but Lua does not require compilation, so both have different advantages.

I think it is very advantageous to have both versions available for use, as long as there are no incompatibilities between the two.

Documentation in /docs is now outdated, and in the doc comments are outdated/broken too.