system_data refactor
* iterate_components and get_components were removed:

I was benchmarking this library and the performance was just garbage ;-;
to optimize this library, I added the untracked field in entity and the
untracked_ids argument on update.

It is also necessary to detect invalid component indexes externally.

So, in these benchmarks, I noticed that using get_components and
iterate_components doubles execution time, so, it don't worth.

the changes are present in example.lua